fix(replays): moves toggle-fullscreen back to controller
Adds a check to make sure the prop is defined
